+++ libs/withdraws/src/index.ts
@@ -1,7 +1,9 @@
 export * from './lib/withdraw-form';
+export * from './lib/withdraw-form-container';
 export * from './lib/withdraw-manager';
 export * from './lib/withdrawals-table';
 export * from './lib/use-complete-withdraw';
 export * from './lib/use-create-withdraw';
 export * from './lib/use-withdrawals';
 export * from './lib/__generated__/Withdrawals';
+export * from './lib/__generated__/WithdrawalFields';
