+++ libs/withdraws/src/lib/withdraw-form.tsx
@@ -19,7 +19,7 @@ import { useWeb3React } from '@web3-react/core';
 import BigNumber from 'bignumber.js';
 import type { ButtonHTMLAttributes, ReactNode } from 'react';
 import { useForm, Controller } from 'react-hook-form';
-import type { WithdrawalFields } from './use-create-withdraw';
+import type { WithdrawalArgs } from './use-create-withdraw';
 import { WithdrawLimits } from './withdraw-limits';
 
 interface FormFields {
@@ -40,7 +40,7 @@ export interface WithdrawFormProps {
     max: BigNumber;
   } | null;
   onSelectAsset: (assetId: string) => void;
-  submitWithdraw: (withdrawal: WithdrawalFields) => void;
+  submitWithdraw: (withdrawal: WithdrawalArgs) => void;
 }
 
 export const WithdrawForm = ({
